requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
huggingface-hub
pypdf
aiofiles
fastapi
uvicorn
jinja2
python-multipart
faiss-cpu
python-dotenv
